backend: setup basic express on app.js + deprecate the boilerplate express
