+++ backend/bin/www
@@ -4,7 +4,7 @@
  * Module dependencies.
  */
 
-var app = require('../app');
+var app = require('../app2');
 var debug = require('debug')('backend:server');
 var http = require('http');
